[FEAT] ProfileService 병합 (#266)
